Get rid of powf, sinf, cosf, fmodf, fabsf, fmaxf, tanf, and roundf

+++ ouzel/audio/Effects.cpp
@@ -56,7 +56,7 @@ namespace ouzel
             void setGain(float newGain)
             {
                 gain = newGain;
-                gainFactor = powf(10.0F, gain / 20.0F);
+                gainFactor = pow(10.0F, gain / 20.0F);
             }
 
         private:
