+++ ouzel/audio/Oscillator.cpp
@@ -73,16 +73,16 @@ namespace ouzel
                 switch (type)
                 {
                     case Oscillator::Type::SINE:
-                        samples[i] = sinf(t * tau<float>());
+                        samples[i] = sin(t * tau<float>());
                         break;
                     case Oscillator::Type::SQUARE:
-                        samples[i] = fmodf(roundf(t * 2.0F + 0.5F), 2.0F) * 2.0F - 1.0F;
+                        samples[i] = fmod(round(t * 2.0F + 0.5F), 2.0F) * 2.0F - 1.0F;
                         break;
                     case Oscillator::Type::SAWTOOTH:
-                        samples[i] = fmodf(t + 0.5F, 1.0F) * 2.0F - 1.0F;
+                        samples[i] = fmod(t + 0.5F, 1.0F) * 2.0F - 1.0F;
                         break;
                     case Oscillator::Type::TRIANGLE:
-                        samples[i] = fabsf(fmodf(t + 0.75F, 1.0F) * 2.0F - 1.0F) * 2.0F - 1.0F;
+                        samples[i] = fabs(fmod(t + 0.75F, 1.0F) * 2.0F - 1.0F) * 2.0F - 1.0F;
                         break;
                 }
