+++ ouzel/scene/Sprite.cpp
@@ -275,7 +275,7 @@ namespace ouzel
                         {
                             if (currentAnimation->repeat)
                             {
-                                currentTime = fmodf(currentTime, length);
+                                currentTime = fmod(currentTime, length);
 
                                 std::unique_ptr<AnimationEvent> resetEvent(new AnimationEvent());
                                 resetEvent->type = Event::Type::ANIMATION_RESET;
@@ -471,7 +471,7 @@ namespace ouzel
                     {
                         if (currentAnimation->repeat)
                         {
-                            currentTime = fmodf(currentTime, length);
+                            currentTime = fmod(currentTime, length);
                             break;
                         }
                         else if (std::next(currentAnimation) == animationQueue.end())
